Extract data page components
