+++ site/AboutThisData.scss
@@ -1,0 +1,84 @@
+.wrapper-about-this-data {
+    width: 100%;
+}
+
+.key-info__title {
+    @include h2-bold;
+    margin: 0 0 16px;
+
+    @include sm-up {
+        margin: 0 0 24px;
+    }
+}
+
+.key-info__content {
+    p {
+        margin: 0;
+    }
+
+    ul,
+    ol {
+        padding-left: 1em;
+        margin: 1em 0;
+
+        li {
+            margin-bottom: 1em;
+
+            &:last-child {
+                margin-bottom: 0;
+            }
+        }
+    }
+}
+
+.key-info__expandable-descriptions {
+    p {
+        margin: 8px 0;
+
+        &:first-of-type {
+            margin-top: 0;
+        }
+
+        &:last-of-type {
+            margin-bottom: 0;
+        }
+    }
+
+    a {
+        @include owid-link-60;
+        color: inherit;
+    }
+}
+
+.key-info__key-description + .key-info__expandable-descriptions {
+    margin-top: 24px;
+}
+
+.key-info__learn-more {
+    @include expandable-paragraph__expand-button--full;
+    margin-top: 24px;
+
+    svg {
+        font-size: 0.75em;
+        margin-left: 12px;
+    }
+}
+
+.key-info__right {
+    @include sm-only {
+        margin-top: 48px;
+    }
+
+    @include sm-up {
+        padding-left: 24px;
+        border-left: 1px solid $blue-20;
+    }
+}
+
+.about-this-data__title {
+    @include datapage-section-title;
+
+    @include md-up {
+        margin-bottom: 0;
+    }
+}
