+++ site/owid.scss
@@ -95,9 +95,13 @@
 @import "./gdocs/components/KeyIndicator.scss";
 @import "./gdocs/components/KeyIndicatorCollection.scss";
 @import "./gdocs/components/PillRow.scss";
+@import "./AboutThisData.scss";
 @import "./DataPage.scss";
 @import "./DataPageContent.scss";
 @import "./GrapherImage.scss";
+@import "./KeyDataTable.scss";
+@import "./MetadataSection.scss";
+@import "./TopicTags.scss";
 @import "./detailsOnDemand.scss";
 
 @import "./gdocs/pages/DataInsight.scss";
